+++ webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/controller/AdminController.java
@@ -30,8 +30,9 @@ import org.oransc.ric.portal.dashboard.DashboardConstants;
 import org.oransc.ric.portal.dashboard.DashboardUserManager;
 import org.oransc.ric.portal.dashboard.model.ErrorTransport;
 import org.oransc.ric.portal.dashboard.model.IDashboardResponse;
-import org.oransc.ric.portal.dashboard.model.RicInstanceKeyName;
-import org.oransc.ric.portal.dashboard.model.RicInstanceList;
+import org.oransc.ric.portal.dashboard.model.RicRegion;
+import org.oransc.ric.portal.dashboard.model.RicRegionList;
+import org.oransc.ric.portal.dashboard.model.RicRegionTransport;
 import org.oransc.ric.portal.dashboard.model.SuccessTransport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -73,7 +74,7 @@ public class AdminController {
 	private DashboardUserManager dashboardUserManager;
 
 	@Autowired
-	private RicInstanceList instanceConfig;
+	private RicRegionList instanceConfig;
 
 	@ApiOperation(value = "Gets the Dashboard MANIFEST.MF property Implementation-Version.", response = SuccessTransport.class)
 	@GetMapping(VERSION_METHOD)
@@ -102,12 +103,12 @@ public class AdminController {
 		return dashboardUserManager.getUsers();
 	}
 
-	@ApiOperation(value = "Gets the list of RIC instances.", response = RicInstanceKeyName.class, responseContainer = "List")
+	@ApiOperation(value = "Gets the RIC regions and instances.", response = RicRegion.class, responseContainer = "List")
 	@GetMapping(INSTANCE_METHOD)
 	@Secured({ DashboardConstants.ROLE_ADMIN, DashboardConstants.ROLE_STANDARD })
-	public List<RicInstanceKeyName> getInstances() {
-		logger.debug("getInstances");
-		return instanceConfig.getKeyNameList();
+	public List<RicRegionTransport> getRegionsInstances() {
+		logger.debug("getRegionsInstances");
+		return instanceConfig.getSimpleInstances();
 	}
 
 	@ApiOperation(value = "Gets the kibana metrics URL for the specified app.", response = SuccessTransport.class)